PR changes and a bug fix
