+++ src/typeorm-aurora-data-api-driver.ts
@@ -1,99 +1,9 @@
 // @ts-ignore
 import createDataApiClient from 'data-api-client'
-
-const transformQuery = (query: string, parameters?: any[]): [string, number] => {
-  const quoteCharacters = ['\'', '"']
-
-  let newQueryString = ''
-  let currentQuote = null
-  let numberOfParametersInQueryString = 0
-  let extendedParameterCount = 0
-
-  for (let i = 0; i < query.length; i += 1) {
-    const currentCharacter = query[i]
-    const currentCharacterEscaped = i !== 0 && query[i - 1] === '\\'
-
-    if (currentCharacter === '?' && !currentQuote) {
-      const parameter = parameters![numberOfParametersInQueryString]
-      if (Array.isArray(parameter)) {
-        parameter.forEach((_, index) => {
-          const isLastParameter = index === parameter.length - 1
-          const paramName = `:e_param_${extendedParameterCount}${isLastParameter ? '' : ', '}`
-          extendedParameterCount += 1
-          newQueryString += paramName
-        })
-      } else {
-        const paramName = `:param_${numberOfParametersInQueryString}`
-        numberOfParametersInQueryString += 1
-        newQueryString += paramName
-      }
-
-      continue
-    }
-
-    if (quoteCharacters.includes(currentCharacter) && !currentCharacterEscaped) {
-      if (!currentQuote) {
-        currentQuote = currentCharacter
-      } else if (currentQuote === currentCharacter) {
-        currentQuote = null
-      }
-    }
-
-    newQueryString += currentCharacter
-  }
-
-  return [newQueryString, numberOfParametersInQueryString]
-}
-
-const transformParameters = (numberOfParametersInQueryString: number, parameters?: any[]) => {
-  if (
-    parameters &&
-    parameters.length > 0 &&
-    parameters.length % numberOfParametersInQueryString !== 0
-  ) {
-    throw new Error(
-      `Number of parameters mismatch, got ${numberOfParametersInQueryString} in query string \
-      and ${parameters.length} in input`)
-  }
-
-  let extendedParameterCount = 0
-  const transformedParameters: any[] = []
-
-  if (parameters && parameters.length > 0) {
-    const numberOfObjects = parameters.length / numberOfParametersInQueryString
-
-    for (let i = 0; i < numberOfObjects; i += 1) {
-      const parameterObject: any = {}
-
-      for (let y = 0; y < numberOfParametersInQueryString; y += 1) {
-        const parameter = parameters[i + y]
-
-        Array.isArray(parameter)
-          // tslint:disable-next-line: no-increment-decrement
-          ? parameter.forEach(element => parameterObject[`e_param_${extendedParameterCount++}`] = element)
-          : parameterObject[`param_${y}`] = parameter
-      }
-
-      transformedParameters.push(parameterObject)
-    }
-  }
-
-  return transformedParameters
-}
+import { transformQueryAndParameters } from './transform.utils'
 
 export default class DataApiDriver {
-  public static transformQueryAndParameters(query: string, parameters?: any[]): any {
-    const [queryString, numberOfParametersInQueryString] = transformQuery(query, parameters)
-    const transformedParameters = transformParameters(numberOfParametersInQueryString, parameters)
-
-    return {
-      queryString,
-      parameters: transformedParameters,
-    }
-  }
-
   private readonly client: any
-
   private transactionId?: string
 
   constructor(
@@ -119,7 +29,7 @@ export default class DataApiDriver {
   }
 
   public async query(query: string, parameters?: any[]): Promise<any> {
-    const transformedQueryData = DataApiDriver.transformQueryAndParameters(query, parameters)
+    const transformedQueryData = transformQueryAndParameters(query, parameters)
 
     this.loggerFn(transformedQueryData.queryString, transformedQueryData.parameters)
