+++ test/functional/utils/create-connection.ts
@@ -30,12 +30,10 @@ export const createConnection = async (partialOptions: Partial<ConnectionOptions
 export const createConnectionAndResetData = async (
   partialOptions: Partial<ConnectionOptions> = {},
 ) => {
-  // FIXME: Whatever happens here the data is not dropped!
-  // Tried: await connection.synchronize(true)
-  // Tried: await connection.dropDatase()
-  // And the following synchronize/dropSchema combination
-  // Bizarre - so for now deleting the data in the tests as need to move on....
-  return createConnection({ ...partialOptions, synchronize: true, dropSchema: true })
+  const connection = await createConnection({ ...partialOptions, synchronize: false })
+  await connection.query('DROP TABLE ALL')
+  await connection.synchronize(true)
+  return connection
 }
 
 export const useCleanDatabase = async (
